agentic_code_analyzer/requirements.txt
google-generativeai
demjson3
google-adk